+++ R/friedman_conover_test.R
@@ -56,6 +56,8 @@ NULL
 #'  freedom. This is the Conover (1999) post-hoc, also known as the
 #'  Durbin-Conover test.
 #'
+#'  The results match \code{PMCMRplus::frdAllPairsConoverTest()}.
+#'
 #'@references Conover, W. J. (1999) Practical Nonparametric Statistics, 3rd
 #'  edition. Wiley.
 #' @seealso \code{\link{friedman_test}}, \code{\link{friedman_nemenyi_test}},